Campaign handling except for notifications

+++ app/Http/Controllers/CampaignController.php
@@ -6,9 +6,12 @@ use Inertia\Inertia;
 use Inertia\Response;
 use App\Models\Member;
 use App\Models\Campaign;
+use App\Models\CampaignRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Redirect;
 use App\Http\Requests\CampaignUpsertRequest;
+use App\Http\Requests\CampaignMembersRequest;
+use Illuminate\Validation\ValidationException;
 
 class CampaignController extends Controller
 {
@@ -59,8 +62,43 @@ class CampaignController extends Controller
 
     public function destroy(Campaign $campaign): RedirectResponse
     {
+        if($campaign->requests()->whereNotNull('ledger_id')->count()) {
+            throw ValidationException::withMessages(['campaign' => 'Campaign has payment activity']);
+        }
         $campaign->delete();
         return Redirect::route('campaign.index')
                        ->with('success', 'Campaign deleted');
+    }
+
+    public function newRequest(CampaignMembersRequest $request, Campaign $campaign )
+    {
+        foreach($request->validated('members') as $userId) {
+            CampaignRequest::create([
+                'campaign_id' => $campaign->id,
+                'user_id' => $userId,
+            ]);
+            // @todo send notification
+        }
+        return Redirect::route('campaign.index')
+                       ->with('success', 'Requests created');
+    }
+
+    public function remindRequest(CampaignMembersRequest $request, Campaign $campaign )
+    {
+        foreach($request->validated('members') as $userId) {
+            // @todo send notification
+        }
+        return Redirect::route('campaign.index')
+                       ->with('success', 'Requests created');
+    }
+
+    public function deleteRequest(CampaignMembersRequest $request, Campaign $campaign )
+    {
+        CampaignRequest::whereIn('user_id', $request->validated('members'))
+                       ->where('campaign_id', $campaign->id)
+                       ->delete();
+
+        return Redirect::route('campaign.index')
+                       ->with('success', 'Requests deleted');
     }
 }
